+++ Assets/Scenes/Playground.unity
@@ -590,6 +590,10 @@ PrefabInstance:
       propertyPath: lowbar
       value: 
       objectReference: {fileID: 1655516952}
+    - target: {fileID: 3153332758830519599, guid: 5cdcb3735fe159e4aa65f3f6a71533d9, type: 3}
+      propertyPath: m_BaseOffset
+      value: 0.7
+      objectReference: {fileID: 0}
     - target: {fileID: 5877653098079884585, guid: 5cdcb3735fe159e4aa65f3f6a71533d9, type: 3}
       propertyPath: m_RootOrder
       value: 9
@@ -638,7 +642,8 @@ PrefabInstance:
       propertyPath: m_Name
       value: Player
       objectReference: {fileID: 0}
-    m_RemovedComponents: []
+    m_RemovedComponents:
+    - {fileID: 5877653098079884588, guid: 5cdcb3735fe159e4aa65f3f6a71533d9, type: 3}
   m_SourcePrefab: {fileID: 100100000, guid: 5cdcb3735fe159e4aa65f3f6a71533d9, type: 3}
 --- !u!1001 &631089227
 PrefabInstance:
