+++ Assets/Scripts/Movement/Gravity.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1e548d60fc8f5cd47bf7d032677cf6f6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
