+++ Assets/Scripts/Movement/Lockable.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3f5dbb6a4e62fe147a12a8e8b929b4b6
+guid: eef89093fc2d4a14eb181c1492612e3b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
